Allow caas to create networks itself
If cluster_network isn't supplied as an extravar, create a network.
Switch to creating ports and associating with instances, and allow
providing cluster_vnic_type and cluster_vnic_profile to enable
direct-mode ports if required. Update variables to use the facts
from the terraform output and not consume cluster_floating_ip
directly.